- use zod to validate data for edit and delete actions on server side   - change arg types for actions to `unknown` since there is possible     uncertainty data that client could provide (use zod validation to     check that data)

+++ src/actions/actions.ts
@@ -1,12 +1,10 @@
 "use server";
 
 import prisma from "@/lib/db";
-import { PetEssentials } from "@/lib/types";
-import { petFormSchema } from "@/lib/validations";
-import { Pet } from "@prisma/client";
+import { petFormSchema, petIdSchema } from "@/lib/validations";
 import { revalidatePath } from "next/cache";
 
-export async function addPet(petData: PetEssentials) {
+export async function addPet(petData: unknown) {
   const validatedPetData = petFormSchema.safeParse(petData);
   if (!validatedPetData.success) {
     return {
@@ -27,13 +25,21 @@ export async function addPet(petData: PetEssentials) {
   revalidatePath("/app", "layout");
 }
 
-export async function editPet(petId: Pet["id"], newPetData: PetEssentials) {
+export async function editPet(petId: unknown, newPetData: unknown) {
+  const validatedPetId = petIdSchema.safeParse(petId);
+  const validatedPetData = petFormSchema.safeParse(newPetData);
+  if (!validatedPetData.success || !validatedPetId.success) {
+    return {
+      message: "Invalid pet data.",
+    };
+  }
+
   try {
     await prisma.pet.update({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
-      data: newPetData,
+      data: validatedPetData.data,
     });
   } catch (error) {
     return {
@@ -44,11 +50,18 @@ export async function editPet(petId: Pet["id"], newPetData: PetEssentials) {
   revalidatePath("/app", "layout");
 }
 
-export async function deletePet(petId: Pet["id"]) {
+export async function deletePet(petId: unknown) {
+  const validatedPetId = petIdSchema.safeParse(petId);
+  if (!validatedPetId.success) {
+    return {
+      message: "Invalid pet data.",
+    };
+  }
+
   try {
     await prisma.pet.delete({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
     });
   } catch (error) {
